Remove BungeeChat references
As this project supports a different platform and has started going in a different direction to BungeeChat, it no longer makes sense for it to be branded as a BungeeChat fork.

I have therefore renamed it to ProxyChat and detached the fork. Copyright notices have been added for appropriate attribution of the remaining BungeeChat work.